+++ js/main.js
@@ -1,48 +1,12 @@
-  const getRandomPositiveInteger = (a, b) => 
-{
-  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
-  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-}
 
-  const getRandomArrayElement = (elements) =>
- {
-  return elements[getRandomPositiveInteger(0, elements.length - 1)];
-};
-
-  const COMMENTS_ID = [];
-
-  const MESSAGES = [
-'Всё отлично!',
-'В целом всё неплохо.Но не всё.',
-'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
-'В конце концов это просто непрофессионально.',
-'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-'Лица у людей на фотке перекошены, как будто их избивают.',
-'Как можно было поймать такой неудачный момент?!',
-]
-
-  const NAMES = [
-'Уилл',
-'Томми',
-'Фил',
-'Алекс',
-'Джеймс',
-'Эш',
-'Марк',
-'Джо',
-'Чарли',
-]
-
+  import '.data.js'
+  import '.util.js'
 
 const COMMENTS = () => {
 const randomIdIndex = getCommentId();
 const randomAvatarIndex = getRandomPositiveInteger(1,6);
 const randomMessageIndex = getRandomArrayElement(MESSAGES);
 const randomNameIndex = getRandomArrayElement(NAMES);
-
 return {
   id: randomIdIndex,
   avatar: 'img/avatar-' + randomAvatarIndex + '.svg',
